# Apps/ThePlaceToMeet.Wpf.App/OpenAPIs/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "ThePlaceToMeet.WebApi",
    "version": "1.0"
  },
  "paths": {
    "/Catering": {
      "get": {
        "tags": [
          "Catering"
        ],
        "operationId": "CateringCaterings",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Catering"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/Catering/{id}": {
      "get": {
        "tags": [
          "Catering"
        ],
        "operationId": "CateringGetBy",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Catering"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/Customer": {
      "get": {
        "tags": [
          "Customer"
        ],
        "operationId": "CustomerCustomers",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Customer"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Customer"
        ],
        "operationId": "CustomerAdd",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Customer"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/Customer/{email}": {
      "get": {
        "tags": [
          "Customer"
        ],
        "operationId": "CustomerGetByEmail",
        "parameters": [
          {
            "name": "email",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Customer"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/MeetingRoom": {
      "get": {
        "tags": [
          "MeetingRoom"
        ],
        "operationId": "MeetingRoomMeetingRooms",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MeetingRoom"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/MeetingRoom/max/{maxAantalPersonen}": {
      "get": {
        "tags": [
          "MeetingRoom"
        ],
        "operationId": "MeetingRoomGetByMaxAantalPersonen",
        "parameters": [
          {
            "name": "maaxNumberOfPersons",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "maxAantalPersonen",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MeetingRoom"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/MeetingRoom/{id}": {
      "get": {
        "tags": [
          "MeetingRoom"
        ],
        "operationId": "MeetingRoomGetById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeetingRoom"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Catering": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "titel": {
            "type": "string",
            "nullable": true
          },
          "beschrijving": {
            "type": "string",
            "nullable": true
          },
          "prijsPerPersoon": {
            "type": "number",
            "format": "double"
          }
        },
        "additionalProperties": false
      },
      "Customer": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "lastName": {
            "type": "string",
            "nullable": true
          },
          "firstName": {
            "type": "string",
            "nullable": true
          },
          "mobile": {
            "type": "string",
            "nullable": true
          },
          "company": {
            "type": "string",
            "nullable": true
          },
          "reservations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Reservation"
            },
            "nullable": true,
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
      "Discount": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "percentage": {
            "type": "integer",
            "format": "int32"
          },
          "minimumReservationsInAYear": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "MeetingRoom": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "naam": {
            "type": "string",
            "nullable": true
          },
          "vergaderruimteType": {
            "$ref": "#/components/schemas/MeetingRoomType"
          },
          "maximumAantalPersonen": {
            "type": "integer",
            "format": "int32"
          },
          "prijsPerUur": {
            "type": "number",
            "format": "double"
          },
          "prijsPerPersoonStandaardCatering": {
            "type": "number",
            "format": "double"
          },
          "reservaties": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Reservation"
            },
            "nullable": true,
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
      "MeetingRoomType": {
        "enum": [
          0,
          1,
          2
        ],
        "type": "integer",
        "format": "int32"
      },
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
      "Reservation": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "aantalPersonen": {
            "type": "integer",
            "format": "int32"
          },
          "dag": {
            "type": "string",
            "format": "date-time"
          },
          "beginUur": {
            "type": "integer",
            "format": "int32"
          },
          "duurInUren": {
            "type": "integer",
            "format": "int32"
          },
          "tot": {
            "type": "integer",
            "format": "int32",
            "readOnly": true
          },
          "prijsPerUur": {
            "type": "number",
            "format": "double"
          },
          "prijsPerPersoonStandaardCatering": {
            "type": "number",
            "format": "double"
          },
          "prijsPerPersoonCatering": {
            "type": "number",
            "format": "double"
          },
          "catering": {
            "$ref": "#/components/schemas/Catering"
          },
          "discount": {
            "$ref": "#/components/schemas/Discount"
          },
          "meetingRoom": {
            "$ref": "#/components/schemas/MeetingRoom"
          }
        },
        "additionalProperties": false
      }
    }
  }
}