# Libraries/ThePlaceToMeet.Contracts/DTO/Discount.cs
namespace ThePlaceToMeet.Contracts.DTO
{
    public class Discount
    {
        public int Id { get; set; }
        public int Percentage { get; set; }
        public int MinimumReservationsInAYear { get; set; }
    }
}